feat: removed public comments
